@@ -25,10 +25,10 @@
         }
     ],
     "replace": {
-        "easyrdf/easyrdf":"1.0.*|1.1.*"
+        "easyrdf/easyrdf":"1.1.*"
     },
     "require": {
-        "php": "^7.1|^8.0",
+        "php": "^8.0",
         "ext-dom" : "*",
         "ext-mbstring": "*",
         "ext-pcre": "*",
@@ -38,7 +38,7 @@
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
         "ml/json-ld": "^1.0",
-        "phpunit/phpunit": "^7.5|^8.5|^9.5",
+        "phpunit/phpunit": "^9.5.0|^10.0.0",
         "phpstan/phpstan": "^1.0",
         "phpstan/phpstan-phpunit": "^1.0",
         "semsol/arc2": "^2.4",
