@@ -45,9 +45,9 @@ use Test\TestCase;
 class JsonTest extends TestCase
 {
     /** @var Json */
-    protected $parser = null;
+    protected $parser;
     /** @var Graph */
-    protected $graph = null;
+    protected $graph;
 
     protected function setUp(): void
     {
