@@ -46,10 +46,10 @@ use Test\TestCase;
 class JsonLdTest extends TestCase
 {
     /** @var JsonLd */
-    protected $serialiser = null;
+    protected $serialiser;
 
     /** @var Graph */
-    protected $graph = null;
+    protected $graph;
 
     protected function setUp(): void
     {
