@@ -86,14 +86,11 @@ class NtriplesArray extends Ntriples
         foreach ($graph->toRdfPhp() as $resource => $properties) {
             foreach ($properties as $property => $values) {
                 foreach ($values as $value) {
-                    array_push(
-                        $triples,
-                        [
-                            's' => $this->serialiseResource($resource),
-                            'p' => '<'.$this->escapeString($property).'>',
-                            'o' => $this->serialiseValue($value),
-                        ]
-                    );
+                    $triples[] = [
+                        's' => $this->serialiseResource($resource),
+                        'p' => '<'.$this->escapeString($property).'>',
+                        'o' => $this->serialiseValue($value),
+                    ];
                 }
             }
         }
