@@ -75,7 +75,7 @@ class SerialiseTest extends TestCase
         );
         $this->assertStringContainsString('<title>EasyRdf Serialiser Example</title>', $output);
 
-        if (version_compare(PHP_VERSION, '8.1.0', '>=')) {
+        if (version_compare(\PHP_VERSION, '8.1.0', '>=')) {
             $this->assertStringContainsString('value&#039; =&gt; &#039;http://xmlns.com/foaf/0.1/Person', $output);
         } else {
             $this->assertStringContainsString("'value' =&gt; 'http://xmlns.com/foaf/0.1/Person',", $output);
